requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
cssutils==2.3.0
Django==3.2.4
django-inlinecss==0.3.0
django-tenants==3.3.2
djangorestframework==3.12.4
future==0.18.2
psycopg2==2.9.1
pynliner==0.8.0
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1